api added, jwt added, autentication and authorization added

+++ controllers/api/v1/post_api.js
@@ -24,13 +24,23 @@ module.exports.destroy = async function(req, res){
 
     try{
         let post = await Post.findById(req.params.id);
-        post.remove();
-        
-        await Comment.deleteMany({post: req.params.id});
 
-        return res.status(200).json({
-            message: "Post and associated comments deleted"
-        });
+        if(req.user.id == post.user)
+        {
+            post.remove();
+        
+            await Comment.deleteMany({post: req.params.id});
+
+            return res.status(200).json({
+                message: "Post and associated comments deleted"
+            });
+        }
+        else{
+            return res.status(401).json({
+                message: 'You can not delete this post'
+            })
+        }
+        
     }catch(err){
         console.log('**********', err);
         return res.status(500).json({
